% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar.R
\name{bar_plot}
\alias{bar_plot}
\title{helper function to make ois bar plot in ggplot2}
\usage{
bar_plot(.data, y_as, percent = TRUE, ...)
}
\arguments{
\item{.data}{het dataframe wat wordt meegeven aan de plot}

\item{y_as}{de waarde die op de y-as worden getoond}
}
\description{
helper function to make ois bar plot in ggplot2
}
