% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donut.R
\name{donut_plot}
\alias{donut_plot}
\title{helper function to make ois donut plot in ggplot2}
\usage{
donut_plot(.data, fill, facet = NULL)
}
\arguments{
\item{.data}{het dataframe wat wordt meegeven aan de plot}

\item{fill}{de waarde waarmee de balken worden gevuld}

\item{facet}{de waarde die de groepen onderverdeelt in 'facets'}
}
\description{
helper function to make ois donut plot in ggplot2
}
