% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gray_shades.R
\name{gray_shades}
\alias{gray_shades}
\title{Helper voor kleuren brewer voor OIS huisstijl die grijswaarde bepaald}
\usage{
gray_shades(.data, fill, grijze_cat = NA, dim_output = FALSE, ...)
}
\arguments{
\item{.data}{het dataframe wat wordt meegeven aan de plot}

\item{fill}{de waarde waarmee de balken worden gevuld}

\item{grijze_cat}{steekwoorden die je als grijs wil tonen}
}
\description{
Helper voor kleuren brewer voor OIS huisstijl die grijswaarde bepaald
}
