% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kleuren.R
\name{kleur_fun}
\alias{kleur_fun}
\title{kleuren brewer voor OIS huisstijl}
\usage{
kleur_fun(dim, pal = kleuren$blauw)
}
\arguments{
\item{dim}{het aantal unieke waardes in de fill}
}
\description{
kleuren brewer voor OIS huisstijl
}
