% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelPercentage.R
\name{label_percentage}
\alias{label_percentage}
\title{labels voor percentages in OIS huisstijl}
\usage{
label_percentage(digit)
}
\arguments{
\item{digit}{de waardes op de x-as van 0 t/m 1}
}
\description{
labels voor percentages in OIS huisstijl
}
